Move transpoder code to trxvu module
